fix(coderd/x/nats): address cluster review comments
